// lib/types/components/avatar.d.ts
import { type TColors } from "../generic";

export type TAvatarVariant = "rounded" | "square" | "circle";
export type TAvatarSize = "xs" | "sm" | "md" | "lg" | "xl" | "xxl";
export type TAvatarColors = TColors;
export type TAvatarClassname = string;
export type TAvatarBordered = boolean;
export type TAvatarBorderColor = TColors;